Final_project/tests/requirements.txt
allure-pytest~=2.8.40
pytest~=6.2.2
pytest-xdist~=2.2.1
selenium~=3.141.0
webdriver-manager~=3.3.0
pytest-lazy-fixture~=0.6.3
requests~=2.25.1
Faker~=8.1.0
pymysql~=1.0.2
sqlalchemy~=1.4.13
sqlalchemy-utils~=0.37.2
furl~=2.1.2
selenium-wire~=4.3.0
docker~=5.0.0